remove the int cast

+++ src/main/java/com/neuronrobotics/sdk/addons/kinematics/PidPrismaticLink.java
@@ -38,7 +38,7 @@ public class PidPrismaticLink extends AbstractPrismaticLink{
 	 */
 	@Override
 	public void cacheTargetValueDevice() {
-		channel.setCachedTargetValue((int)getTargetValue());
+		channel.setCachedTargetValue((float)getTargetValue());
 	}
 
 	/* (non-Javadoc)
